Refactor: Связи услуг-запчастей - упрощенная настройка и выбор
- Удалена вкладка 'Общие настройки' из окна Настройки
- Переработана вкладка 'Связи' - убрана форма добавления, оставлена таблица с кнопками
- Удален EditServiceSparePartDialog
- Удален метод updateServiceSparePart из всех классов
- Переработан CreateOrderDialog/EditOrderDialog - выбор запчастей через чекбоксы при добавлении услуги
- Исправлен AutoAddSparePartService - теперь правильно загружает связи из БД
- Добавлен диалог showAddServiceSparePartDialog для создания связей с возможностью указать количество и ед.изм. для каждой запчасти
- Обновлена таблица spare_parts - добавлены unit_type, min_stock, location
- Структура базы данных обновлена для поддержки новых полей
- Улучшен UI настроек и диалогов работы с заказами
- Обновлены тесты и стили

+++ src/main/java/com/autoservice/WorkOrder.java
@@ -17,7 +17,7 @@ public class WorkOrder {
     private List<String> services = new ArrayList<>();
     private List<Double> servicePrices = new ArrayList<>();
     private List<SparePart> spareParts = new ArrayList<>();
-    private List<Integer> sparePartQuantities = new ArrayList<>();
+    private List<Double> sparePartQuantities = new ArrayList<>();
     private boolean dirty = false;
 
     // ==================== КОНСТРУКТОРЫ ====================
@@ -55,7 +55,7 @@ public class WorkOrder {
     public List<String> getServices() { return services; }
     public List<Double> getServicePrices() { return servicePrices; }
     public List<SparePart> getSpareParts() { return spareParts; }
-    public List<Integer> getSparePartQuantities() { return sparePartQuantities; }
+    public List<Double> getSparePartQuantities() { return sparePartQuantities; }
     public boolean isDirty() { return dirty; }
 
     // ==================== СЕТТЕРЫ ====================
@@ -107,7 +107,7 @@ public class WorkOrder {
         }
     }
 
-    public void addSparePart(SparePart part, int quantity) {
+    public void addSparePart(SparePart part, double quantity) {
         this.spareParts.add(part);
         this.sparePartQuantities.add(quantity);
         this.dirty = true;
